macOS: Remove QMacRootLevelAutoReleasePool member from QCoreApplication
Tying the pool to the lifetime of  QCoreApplication becomes a problem when
the application is allocated on the heap, for example when Qt is used to
create plugins, as NSAutoreleasePool should be allocated and drained in
the same context, typically a function scope.

The feature is disabled for now, until we can find a better way to
achieve the same thing.

Change-Id: I600587df62a8cc7caded5172ad914c0e04d794d9
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>